Move graph data to a separate module
